// src/types/api/githubComIntocodePromptarchiveInternalServiceTagResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * PromptArchive API
 * REST API for storing and managing prompts (Go + Gin + PostgreSQL + Redis)
 * OpenAPI spec version: 1.0
 */

export interface GithubComIntocodePromptarchiveInternalServiceTagResponse {
  color?: string;
  id?: string;
  name?: string;
  prompts_count?: number;
}